Got pet calls working but needs to handle 2 states
Also re-named some files

+++ app/src/main/java/com/chaseolson/pets/home/model/PetFinderResponse.kt
@@ -15,7 +15,7 @@ data class PetFinderResponse(
     @Xml
     data class Pet(
             @PropertyElement
-            val id: Long = 0,
+            val id: Int = 0,
             @PropertyElement
             val name: String = "N/A",
             @PropertyElement
